// types/Task.type.ts
export type TaskType = {
  id: number;
  attributes: {
    goal: string;
    createdAt: string;
    updatedAt: string;
    isDone: boolean;
  };
};
// "id": 1,
// "attributes": {
//     "goal": "Define our types",
//     "createdAt": "2024-09-02T21:42:44.430Z",
//     "updatedAt": "2024-09-02T21:42:44.430Z",
//     "isDone": false
//   }